+++ flutter/lib/sentry_flutter.dart
@@ -18,3 +18,4 @@ export 'src/user_interaction/sentry_user_interaction_widget.dart';
 export 'src/binding_wrapper.dart';
 export 'src/sentry_widget.dart';
 export 'src/navigation/sentry_display_widget.dart';
+export 'src/feedback/sentry_feedback_widget.dart';